requirements.txt
python-telegram-bot
PyYAML
